refactor: moved the DEFAULT_QUERY_TEMPLATE to existing file instead of creating new
