+++ studio/src/lib/constants.ts
@@ -127,3 +127,33 @@ export const SHARE_OPTIONS = [
   { id: "preOperation", label: "Pre-request Script", isChecked: false, isDisabled: false },
   { id: "postOperation", label: "Post-request Script", isChecked: false, isDisabled: false },
 ] as const;
+
+export const DEFAULT_QUERY_TEMPLATE = `# Welcome to WunderGraph Studio
+#
+#
+# Type queries into this side of the screen, and you will see intelligent
+# typeaheads aware of the current GraphQL type schema and live syntax and
+# validation errors highlighted within the text.
+#
+# GraphQL queries typically start with a "{" character. Lines that start
+# with a # are ignored.
+#
+# An example GraphQL query might look like:
+#
+#     {
+#       field(arg: "value") {
+#         subField
+#       }
+#     }
+#
+# Keyboard shortcuts:
+#
+#   Prettify query:  Shift-Ctrl-P (or press the prettify button)
+#
+#  Merge fragments:  Shift-Ctrl-M (or press the merge button)
+#
+#        Run Query:  Ctrl-Enter (or press the play button)
+#
+#    Auto Complete:  Ctrl-Space (or just start typing)
+#
+`;
